Capstone Project Part 1
In Part 1 of the capstone project, was developed an ETL data pipeline for DeFtunes, a music streaming company, to handle their new digital song purchase feature. The pipeline was prepared and served data to the data analytics team for analysis.  The pipeline follow a medallion architecture, storing data in an S3 bucket, transforming it using Iceberg tables, and modeling it into a star schema in Redshift using dbt.